# Add in optimizations for softmax for Fusion F1.