Confirmed that the test passes with:
```
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=fusion_f1 XTENSA_CORE=F1_190305_swupgrade test_kernel_softmax_test -j8
```

However, the latency improvement is only ~1000 ticks, as tested with:
```
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=fusion_f1 XTENSA_CORE=F1_190305_swupgrade test_keyword_benchmark -j8
```

With the optimized implementation:
```
InitializeKeywordRunner() took 205971 ticks (205 ms)
KeywordRunNIerations(1) took 150995 ticks (150 ms)
KeywordRunNIerations(10) took 1509457 ticks (1509 ms)
```

and with the reference implementation:
```
InitializeKeywordRunner() took 201464 ticks (201 ms)
KeywordRunNIerations(1) took 152158 ticks (152 ms)
KeywordRunNIerations(10) took 1521087 ticks (1521 ms)
```