@advaitjain 2021-01-28
@pnikam-cad @nyadla-sys: The kernel tests pass ok with the optimized implementation. However, I am not seeing any meaningful latency improvement.

This could happen for any combination of the following reasons:
 * the optimizations could be improved for the case of int8 inputs and int16 output.
 * for the Fusion F1, the reference implementation of softmax doesn't really have much room to be optimized.
 * for the fusion F1 (in contrast to the hifimini), the contribution of softmax to the overall latency is minimal.

I'll look more closely at these tomorrow but sharing the current status with you guys.