@pnikam-cad 2021-02-03
I looked a little bit into the softmax performance. Couple of observations, inline with the points mentioned above :
- The reference implementation is already well optimized because of the lookup table approach, and the contribution of the softmax operator is much less now. So we may not see big difference when switching to optimized implementation.
- For the softmax kernel in the HiFi 4 NN library, this case is going into a slightly unoptimized path compared to HiFi mini. We should be able to fix this in the next version, that should give us some improvement.