@advaitjain 2021-02-12
I have updated the commit description with the latest per-op profiling. Especially with all the other optimizations, the ~1500 ticks improvement that we get from softmax is worth merging.